implementation book_a_cab_at, journey complete! , displacement and billing calculations

+++ db/migrate/20180517044236_create_cabs.rb
@@ -1,8 +1,8 @@
 class CreateCabs < ActiveRecord::Migration[5.1]
   def change
     create_table :cabs do |t|
-      t.decimal :position_x
-      t.decimal :position_y
+      t.float :position_x
+      t.float :position_y
       t.string :kind
 
       t.timestamps
